separate observers impl from wallet proxy
